+++ internal/gateway/gateway_definition.go
@@ -1,7 +1,9 @@
 package gateway
 
 import (
+	"errors"
 	"fmt"
+	"time"
 
 	"github.com/go-logr/logr"
 
@@ -16,12 +18,12 @@ func gatewayDefinitionCreated(
 	r *controller.Reconciler,
 	gatewayDefinition *v0.GatewayDefinition,
 	log *logr.Logger,
-) error {
+) (int64, error) {
 
 	// create gateway kubernetes manifests
 	yamlDocument, err := createYAMLDocument(r, gatewayDefinition)
 	if err != nil {
-		return fmt.Errorf("failed to create yaml document: %w", err)
+		return 0, fmt.Errorf("failed to create yaml document: %w", err)
 	}
 
 	// construct workload definition object
@@ -35,7 +37,7 @@ func gatewayDefinitionCreated(
 	// create workload definition
 	createdWorkloadDefinition, err := client.CreateWorkloadDefinition(r.APIClient, r.APIServer, &workloadDefinition)
 	if err != nil {
-		return fmt.Errorf("failed to create workload definition in threeport API: %w", err)
+		return 0, fmt.Errorf("failed to create workload definition in threeport API: %w", err)
 	}
 
 	// update gateway definition
@@ -48,7 +50,7 @@ func gatewayDefinitionCreated(
 		gatewayDefinition,
 	)
 	if err != nil {
-		return fmt.Errorf("failed to update gateway definition in threeport API: %w", err)
+		return 0, fmt.Errorf("failed to update gateway definition in threeport API: %w", err)
 	}
 
 	log.V(1).Info(
@@ -56,7 +58,7 @@ func gatewayDefinitionCreated(
 		"gatewayDefinitionID", gatewayDefinition.ID,
 	)
 
-	return nil
+	return 0, nil
 }
 
 // gatewayDefinitionUpdated performs reconciliation when a gateway definition
@@ -65,17 +67,17 @@ func gatewayDefinitionUpdated(
 	r *controller.Reconciler,
 	gatewayDefinition *v0.GatewayDefinition,
 	log *logr.Logger,
-) error {
+) (int64, error) {
 
 	// create gateway kubernetes manifests
 	yamlDocument, err := createYAMLDocument(r, gatewayDefinition)
 	if err != nil {
-		return fmt.Errorf("failed to create yaml document: %w", err)
+		return 0, fmt.Errorf("failed to create yaml document: %w", err)
 	}
 
 	// get workload definition
 	if gatewayDefinition.WorkloadDefinitionID == nil {
-		return fmt.Errorf("failed to update workload definition, workload definition ID is nil")
+		return 0, fmt.Errorf("failed to update workload definition, workload definition ID is nil")
 	}
 	workloadDefinition, err := client.GetWorkloadDefinitionByID(
 		r.APIClient,
@@ -83,14 +85,14 @@ func gatewayDefinitionUpdated(
 		*gatewayDefinition.WorkloadDefinitionID,
 	)
 	if err != nil {
-		return fmt.Errorf("failed to get workload definition by workload definition ID: %w", err)
+		return 0, fmt.Errorf("failed to get workload definition by workload definition ID: %w", err)
 	}
 
 	// update workload definition
 	workloadDefinition.YAMLDocument = &yamlDocument
 	_, err = client.UpdateWorkloadDefinition(r.APIClient, r.APIServer, workloadDefinition)
 	if err != nil {
-		return fmt.Errorf("failed to update workload definition in threeport API: %w", err)
+		return 0, fmt.Errorf("failed to update workload definition in threeport API: %w", err)
 	}
 
 	// update gateway definition
@@ -103,7 +105,7 @@ func gatewayDefinitionUpdated(
 		gatewayDefinition,
 	)
 	if err != nil {
-		return fmt.Errorf("failed to update gateway definition in threeport API: %w", err)
+		return 0, fmt.Errorf("failed to update gateway definition in threeport API: %w", err)
 	}
 
 	log.V(1).Info(
@@ -111,7 +113,7 @@ func gatewayDefinitionUpdated(
 		"gatewayDefinitionID", workloadDefinition.ID,
 	)
 
-	return nil
+	return 0, nil
 }
 
 // gatewayDefinitionDeleted performs reconciliation when a gateway definition
@@ -120,15 +122,55 @@ func gatewayDefinitionDeleted(
 	r *controller.Reconciler,
 	gatewayDefinition *v0.GatewayDefinition,
 	log *logr.Logger,
-) error {
+) (int64, error) {
+	// check that deletion is scheduled - if not there's a problem
+	if gatewayDefinition.DeletionScheduled == nil {
+		return 0, errors.New("deletion notification receieved but not scheduled")
+	}
+
+	// check to see if reconciled - it should not be, but if so we should do no
+	// more
+	if gatewayDefinition.DeletionConfirmed != nil {
+		return 0, nil
+	}
 
 	// delete workload definition
 	if gatewayDefinition.WorkloadDefinitionID == nil {
-		return fmt.Errorf("failed to delete workload definition, workload definition ID is nil")
+		return 0, fmt.Errorf("failed to delete workload definition, workload definition ID is nil")
 	}
 	_, err := client.DeleteWorkloadDefinition(r.APIClient, r.APIServer, *gatewayDefinition.WorkloadDefinitionID)
 	if err != nil {
-		return fmt.Errorf("failed to delete workload definition with ID %d: %w", *gatewayDefinition.WorkloadDefinitionID, err)
+		return 0, fmt.Errorf("failed to delete workload definition with ID %d: %w", *gatewayDefinition.WorkloadDefinitionID, err)
+	}
+
+	// delete the gateway definition that was scheduled for deletion
+	deletionReconciled := true
+	deletionTimestamp := time.Now().UTC()
+	deletedGatewayDefinition := v0.GatewayDefinition{
+		Common: v0.Common{
+			ID: gatewayDefinition.ID,
+		},
+		Reconciliation: v0.Reconciliation{
+			Reconciled:           &deletionReconciled,
+			DeletionAcknowledged: &deletionTimestamp,
+			DeletionConfirmed:    &deletionTimestamp,
+		},
+	}
+	_, err = client.UpdateGatewayDefinition(
+		r.APIClient,
+		r.APIServer,
+		&deletedGatewayDefinition,
+	)
+	if err != nil {
+		return 0, fmt.Errorf("failed to confirm deletion of gateway definition in threeport API: %w", err)
+	}
+	_, err = client.DeleteGatewayDefinition(
+		r.APIClient,
+		r.APIServer,
+		*gatewayDefinition.ID,
+	)
+	if err != nil {
+		return 0, fmt.Errorf("failed to delete gateway definition in threeport API: %w", err)
 	}
 
 	log.V(1).Info(
@@ -136,7 +178,7 @@ func gatewayDefinitionDeleted(
 		"gatewayDefinitionID", gatewayDefinition.ID,
 	)
 
-	return nil
+	return 0, nil
 }
 
 // createYAMLDocument creates a YAML document containing the Kubernetes
